requirements.txt
discord.py>=2.3.2
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
yfinance>=0.2.36
plotly>=5.20.0

